Rerouted scratchpad and added to main schematic
Rerouted the scratchpad traces so that it has only 9 common pins, and two sets of anodes 
A-H and A-P for 16 segment, where each common goes to half a 16 seg digit, and 7 segments respectively, where each common goes to two 7 seg digits. I plan to tie the first set of A-H together with the second A-H, but the option to separately multiplex them is there.
Also replaced the 1mm pitch headers with 2.54mm headers
Added a symbol to the library and added it to the main UFC schematic.