Split up board, started routing traces
Moved the LED drivers and the STM32 board off of the main board and onto a daughterboard to fit a reasonable space. Second board will go behind the first, connected with .1" headers. 
Additionally, traces on the bottom four (second LED driver) OPT displays are routed. 
Redid the OPT buttons to be included in the keypad matrix, to save pins.
Removed the shift register pins and wired it all directly to the STM32, to save space