Finished routing, updated comm channel displays
Decided to use (and again modify) open hornet's com channel indicators. Finished routing traces and adding ground planes on the main board, though it may be a little sloppy. Began work on the back side board, which will have the LED drivers and STM32. Re added the shift register board for now for the sake of having more pins.